requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-filter==23.5
python-decouple==3.8
psycopg2-binary==2.9.9
drf-yasg==1.21.7
coverage==7.3.2
Pillow==10.1.0

